fastapi_app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
jinja2
python-dotenv
# requests # Für HTTP-Anfragen an andere Dienste oder APIs
sqlalchemy